+++ stroom-core-shared/src/main/java/stroom/pipeline/shared/data/PipelineDataUtil.java
@@ -68,26 +68,20 @@ public class PipelineDataUtil {
         return property;
     }
 
-    public static PipelineReference createReference(final String element, final String name,
-                                                    final DocRef pipeline, final DocRef feed, final String streamType) {
-        final PipelineReference pipelineReference = new PipelineReference();
-        pipelineReference.setElement(element);
-        pipelineReference.setName(name);
-        pipelineReference.setPipeline(pipeline);
-        pipelineReference.setFeed(feed);
-        pipelineReference.setStreamType(streamType);
-        return pipelineReference;
+    public static PipelineReference createReference(final String element,
+                                                    final String name,
+                                                    final DocRef pipeline,
+                                                    final DocRef feed,
+                                                    final String streamType) {
+        return new PipelineReference(element, name, pipeline, feed, streamType);
     }
 
-    public static PipelineReference createReference(final String element, final String name,
-            final PipelineEntity pipeline, final Feed feed, final String streamType) {
-        final PipelineReference pipelineReference = new PipelineReference();
-        pipelineReference.setElement(element);
-        pipelineReference.setName(name);
-        pipelineReference.setPipeline(DocRef.create(pipeline));
-        pipelineReference.setFeed(DocRef.create(feed));
-        pipelineReference.setStreamType(streamType);
-        return pipelineReference;
+    public static PipelineReference createReference(final String element,
+                                                    final String name,
+                                                    final PipelineEntity pipeline,
+                                                    final Feed feed,
+                                                    final String streamType) {
+        return new PipelineReference(element, name, DocRef.create(pipeline), DocRef.create(feed), streamType);
     }
 
     public static PipelineLink createLink(final String from, final String to) {