+++ stroom-pipeline/src/main/java/stroom/refdata/ReferenceData.java
@@ -24,7 +24,6 @@ import stroom.feed.shared.Feed;
 import stroom.pipeline.shared.data.PipelineReference;
 import stroom.pipeline.state.FeedHolder;
 import stroom.pipeline.state.StreamHolder;
-import stroom.security.shared.DocumentPermissionNames;
 import stroom.streamstore.server.fs.serializable.StreamSourceInputStream;
 import stroom.streamstore.server.fs.serializable.StreamSourceInputStreamProvider;
 import stroom.streamstore.shared.Stream;
@@ -313,10 +312,12 @@ public class ReferenceData {
     private static class CachedMapStore {
         private final long streamNo;
         private final MapStore mapStore;
+        private final int hashCode;
 
         CachedMapStore(final long streamNo, final MapStore mapStore) {
             this.streamNo = streamNo;
             this.mapStore = mapStore;
+            hashCode = Long.hashCode(streamNo);
         }
 
         public long getStreamNo() {
@@ -328,19 +329,16 @@ public class ReferenceData {
         }
 
         @Override
-        public int hashCode() {
-            return (int) streamNo;
+        public boolean equals(final Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+            final CachedMapStore that = (CachedMapStore) o;
+            return streamNo == that.streamNo;
         }
 
         @Override
-        public boolean equals(final Object obj) {
-            if (obj == null || !(obj instanceof CachedMapStore)) {
-                return false;
-            }
-
-            final CachedMapStore cachedMapStore = (CachedMapStore) obj;
-
-            return cachedMapStore.streamNo == streamNo;
+        public int hashCode() {
+            return hashCode;
         }
     }
 }